Add files via upload
This program asks the client for their midterm, final, and homework scores. Requests
for the weight of the work type and can calculates midterm and final scores 
factoring in a user's shift amount. Based off the calculated sum of the client's weighted 
midterm, final, and homework scores, prints out an appropriate GPA and message.